+++ HISTORY.rst
@@ -7,6 +7,13 @@ TO_BE_ANNOUNCED
 
 - list products with audience rating 18+ in notification mail if region is Germany and product is also 18+ `#92 <https://github.com/ponyriders/django-amazon-price-monitor/issues/92>`__ (`PR#93 <https://github.com/ponyriders/django-amazon-price-monitor/pull/93>`__)
 
+`0.6.1 <https://pypi.python.org/pypi/django-amazon-price-monitor/0.6.1>`__
+--------------------------------------------------------------------------
+**Bugfixes:**
+
+- StartupTask fails with exception `#94 <https://github.com/ponyriders/django-amazon-price-monitor/issues/94>`__
+- Tests fail if today is the last day of November `#95 <https://github.com/ponyriders/django-amazon-price-monitor/issues/95>`__
+
 `0.6 <https://pypi.python.org/pypi/django-amazon-price-monitor/0.6>`__
 ----------------------------------------------------------------------
 **Features:**